Add files via upload
I build a model using the Random Forest Classifier. Here the approach I have used is called Landmark Mapping. Where we use MediaPipe framework to project 42 points on each hand to track its movement. I think this approach is easier and more accurate than using a CNN based model. Here the model is very light weight and works at a mediocre level. Main things that we need to work on now would be finding a dataset which has a variety of pictures (huge band spectrum), and to decide upon which classification technique we will be using for the model. Depending on which technique we employ, our model's size would vary and so will the functionality.