Added model.p via upload
This model has been trained on my own dataset, where I captured 100 different images for numbers 0 to 9.  The model is ~95% accurate and if we can add more data I am sure there will be more variety in the dataset, making the model more robust, but this is good enough for testing.